Add docs in generated files
The tbldef structs usually contain a description field which we can use
to print a nice description of the operation. Unfortunately Haddock uses
syntax which is quite different from Markdown, so the translation isn't
perfect, but I don't think that adding a dependency on a full blown
Markdown parser is a good idea.