Add DenseArrayAttr type
DenseArrayAttr's unsigned/signed variants are still being discussed,
mapping all to the signless for now.

Change the generator to add suffic for most args to avoid shadowing
warnings (Shape dialect has an op that has the same name as an attribute
of another op which results in failure).

Also flip cond.br to use it to fix test.